Move NameResolver into common_socket_impl
